+++ src/storage/backup_manager.cpp
@@ -171,7 +171,7 @@ bool isValidCronExpression(const std::string& expression) {
         fields.push_back(field);
     }
 
-    if (static_cast<int>(fields.size()) != 5) {
+    if (fields.size() != 5) {
         return false;
     }
 
@@ -416,7 +416,7 @@ std::optional<RemoteBackupLocation> parseRemoteBackupLocation(StorageBackend bac
     }
     case StorageBackend::AZURE: {
         const auto segments = splitPathSegments(payload);
-        if (static_cast<int>(segments.size()) < 2) {
+        if (segments.size() < 2) {
             return std::nullopt;
         }
 
@@ -838,7 +838,7 @@ RAIDConfig BackupManager::detectRAIDConfiguration() {
             
         case RAIDMode::RAID5:
             // RAID5: N-1 data shards, 1 parity shard
-            if (static_cast<int>(config.shards.size()) >= 3) {
+            if (config.shards.size() >= 3) {
                 config.data_shards = static_cast<uint32_t>(config.shards.size() - 1);
                 config.parity_shards = 1;
                 config.is_coordinated = true;  // Need all shards (data + parity)
@@ -847,7 +847,7 @@ RAIDConfig BackupManager::detectRAIDConfiguration() {
             
         case RAIDMode::RAID6:
             // RAID6: N-2 data shards, 2 parity shards
-            if (static_cast<int>(config.shards.size()) >= 4) {
+            if (config.shards.size() >= 4) {
                 config.data_shards = static_cast<uint32_t>(config.shards.size() - 2);
                 config.parity_shards = 2;
                 config.is_coordinated = true;  // Need all shards (data + double parity)
@@ -2799,10 +2799,10 @@ bool BackupManager::performPITR(const std::string& dest_dir, const PITROptions& 
         for (const auto& backup_name : backups) {
             // Only consider full backups (incremental replay not yet implemented).
             static constexpr std::string_view kPrefix = "full_";
-            if (static_cast<int>(backup_name.size()) < static_cast<int>(kPrefix.size()) + 15) {
+            if (backup_name.size() < kPrefix.size() + 15) {
               continue;
             }
-            if (backup_name.compare(0,static_cast<int>(kPrefix.size()), kPrefix) != 0) {
+            if (backup_name.compare(0, kPrefix.size(), kPrefix) != 0) {
               continue;
             }
